Added source code for the library
Steve was born in a fork of the CrystalFontz/CFA480128Ex-039Tx repo. I noticed that the CrystalFontz / FTDI / Bridgetek sample code depended on #defines a little too much to my liking. This made it difficult to read through the code and figure out what each macro was for. Also, it made the code type-unsafe: it would be easy to put totally the wrong value into a command or register by mistake, making a program fail silently (without a warning or error from the compiler). Furthermore, I noticed that the macros made it necessary to re-compile the entire source code for every single display, because display parameters (such as width and height) were defined in macros too. Another thing that bothered me a little was that the sample code was Arduino specific.

This inspired me to write a C++ class with code to support the EVE display chips, that fixes those problems. I called it Static Type-safe EVE because the resulting acronym is also a running joke in my family.
* All the "magic values" that are intended for a specific purpose, are declared as enums, for type safety.
* All the values that are specific to a display are stored in a Display Profile class called SteveDisplay. For many displays, an instance of the display class can be created to represent the parameters of the display. For more complicated displays, a subclass can be created with a constructor that changes some parameters from the default.
* The way that the host communicates with the EVE chip is implemented in SteveHAL. For example, when used with Arduino, a program can use the SteveHAL_Arduino class with uses the SPI.h module. If necessary, variations can be created that use QSPI or higher SPI clock frequencies. For other platforms such as Windows with an FTDI 4222H, other HALs can be created.
* At the time of this writing, touch screen code is not implemented yet, but will be added soon. This will be implemented in a separate class, possibly multiple classes depending on which type of touch screen controller is in use. If you don't need touch screen functionality, it won't need to be in your project.
* At the time of this writing, audio is not implemented yet, but will be added soon. This may also be implemented as a separate class.