Created a HAL for use with MPSSE compatible devices such as the FTDI C232HM-DDHSL-0
It kinda works but not really: It generates a frame about every second or so, and freezes after a while. I suspect I need to tweak the D2XX layer a bit to improve throughput, or change the HAL (and possibly Steve itself) so that it sends/receives more data at a time.

Overview:
* Added a Begin and End interface to the HAL to allow intialization early in the Steve.Begin function. Originally the hardware didn't get a chance to initialize until after the CS and PD lines are already changed, which is not going to work of course. Silly me.
* The SteveHAL_Windows_MPSSE module does have a .cpp file but that's only to make Intellisense not go nuts.
* I added property sheet used by the Windows projects (as usual in my bigger Visual Studio projects) to make it easier to change include directories and library directories and other settings for all projects.