Changed MPSSE HAL to use SPI_Read and SPI_Write
Also used buffered writes for all HAL read/write calls instead of writing bytes in a loop.

That changes the loop time of the demo from 1200 ms to 12 ms, but unfortunately something is still not right: Lots of glitches on the screen (CFA480128) and the display stops updating after a few seconds.

Going to try to buffer transfers up between CS on/off commands and see if it makes a diference. This is what the EVE sample program in the MPSSE source tree does.