Implemented a write cache in the MPSSE HAL
Unfortunately even though the loop time is now down to 7 ms or so, this doesn't fix the problem: The screen is still glitchy and the display freezes after a few seconds.

Maybe this is a hardware problem? Going to check with the other display (that has fewer... "antennae" - i.e. self soldered wires) sticking out of the breakout board).

It's also possible that the cable really doesn't have enough power for the CFA480128 display; I should try powering it in a different way.